pyproject.toml
[project]
name = "kite"
version = "0.1.0"
description = "KITE - The Keyword Insight and Term Extraction Project"
authors = [
    {name = "David T Thompson", email = "david.thompson86@gmail.com"},
]
dependencies = ["haystack-ai>=2.12.1", "markdown-it-py>=3.0.0", "mdit-plain>=1.0.1", "nltk>=3.9.1", "sentence-transformers>=3.0.0", "accelerate>=1.6.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.scripts]
lint = { composite = ["black --check .", "isort --check-only src tests"] }
format = { composite = ["black .", "isort src tests"] }


[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]


[tool.isort]
profile = "black"
src_paths = ["src", "tests"]